F2 S2: the barrel shifter, as a gadget, tested before anything uses it
A right shift of a boolean vector by a witnessed amount, which is the one
piece of machinery three queued items all need: a subnormal RESULT in the
multiply, the exponent alignment in an fp32 accumulator, and a subnormal
source in a widening.

One conditional stage per amount bit, applied top down, so stage k shifts
by 2^(bits-1-k) and the stages compose. Two things this IR cannot say
directly, both handled: "i + amount" runs off the top of the vector, so
that branch reads as the constant ZERO rather than a column that is not
there, and OR is an inequality, so the sticky is a prefix-OR chain (one
constraint per bit, x OR y = x + y - x·y, boolean for free) plus a
per-stage flag and an accumulator. 152 constraints per row for a 25-bit
vector with a 4-bit amount, degree 2 throughout.

Five bugs lived in here and all five were in the gadget rather than the
caller, which is the argument for testing it alone first: the stage chain
read the wrong slot, the control bit and the shift distance disagreed,
the prefix columns were allocated in ascending order while the stages
needed descending, the flag read the first prefix column instead of the
OR of the block, and the shift was a left shift of the indices when the
value moves right. The sticky also ended in an internal column instead of
the caller's.

Swept over every shift amount crossed with vectors that hit each case,
every constraint evaluated, no prover. The prove/verify case uses a
smaller configuration on purpose: 8 rows of the 25-bit one is 1216
composed constraints, over the current cap of 1024, and raising that cap
is a decision to make when the multiply actually needs it rather than in
a test.

air_builder's writeBits now takes u64, because an fp32 significand plus
its product needs 49 bits and the u32 parameter could not hold it.

+++ libs/stark/air_builder.zig
@@ -248,8 +248,9 @@ pub const Trace = struct {
 
     /// Bit i of `v` goes to COLUMN (base + i), row r. The trace is
     /// column-major, so a run of bits is a stride over columns, not a
-    /// contiguous range.
-    pub fn writeBits(self: *Trace, base: u16, v: u32, n: u16, r: usize) void {
+    /// contiguous range. `v` is u64 because an fp32 significand plus its
+    /// product needs 49 bits.
+    pub fn writeBits(self: *Trace, base: u16, v: u64, n: u16, r: usize) void {
         for (0..n) |i| {
             self.columns[base + @as(u16, @intCast(i))][r] = Fp2.re(Goldilocks.fromU64((v >> @intCast(i)) & 1));
         }
